+++ webadmin/var/www/webadmin/netBoot.php
@@ -15,26 +15,20 @@ $currentSubnet = trim(getNetAddress($currentIP, $currentNetmask));
 $netbootimgdir = "/srv/NetBoot/NetBootSP0/";
 $subnetcheck = $conf->getSubnets();
 
-if (isset($_POST['netbootName']))
-{
-	$conf->setSetting("netbootname", $_POST['netbootName']);
-}
-
-if ((isset($_POST['enablenetboot']) || isset($_POST['changenetboot'])) && empty($subnetcheck))
+if (isset($_POST['enablenetboot']) && empty($subnetcheck))
 {
 	echo "<div class=\"alert alert-danger\">ERROR: Ensure you added a proper Subnet and Netmask</div>";
 }
 
-if ((isset($_POST['enablenetboot']) || isset($_POST['changenetboot'])) && !isset($_POST['NetBootImage']))
+if (isset($_POST['enablenetboot']) && (!isset($_POST['NetBootImage']) || $_POST['NetBootImage'] == ""))
 {
-	echo "<div class=\"alert alert-danger\">ERROR: Ensure you have uploaded a properly configured NetBoot image</div>";
+	echo "<div class=\"alert alert-danger\">ERROR: Ensure you have uploaded and selected a properly configured NetBoot image</div>";
 }
 
 
-if (isset($_POST['NetBootImage']))
+if (isset($_POST['NetBootImage']) && $_POST['NetBootImage'] != "")
 {
 	$wasrunning = getNetBootStatus();
-	$netbootname = $conf->getSetting("netbootname");
 	$nbi = $_POST['NetBootImage'];
 	if ($nbi != "")
 	{
@@ -54,12 +48,12 @@ if (isset($_POST['NetBootImage']))
 		suExec("disablenetboot");
 		suExec("installdhcpdconf");
 
-		if ($wasrunning || isset($_POST['enablenetboot']) || isset($_POST['changenetboot'])) {
-			suExec("setnbimages " . $nbi . " " . $netbootname);
+		if ($wasrunning || isset($_POST['enablenetboot'])) {
+			suExec("setnbimages " . $nbi);
 		}
 		$conf->setSetting("netbootimage", $nbi);
 
-		if ((isset($_POST['enablenetboot']) || isset($_POST['changenetboot'])) && !getNetBootStatus() && !empty($subnetcheck)) {
+		if (isset($_POST['enablenetboot']) && !getNetBootStatus() && !empty($subnetcheck)) {
 			echo "<div class=\"alert alert-danger\">ERROR: Unable to start NetBoot service. Ensure your .nbi directory is properly configured</div>";
 		}
 	}
@@ -130,16 +124,16 @@ if (isset($_GET['deleteSubnet']) && isset($_GET['deleteNetmask'])
 if (!isset($_POST['disablenetboot']) && getNetBootStatus())
 {
 	$tftp_running = (trim(suExec("gettftpstatus")) === "true");
-	// $nfs_running = (trim(suExec("getnfsstatus")) === "true");
+	$nfs_running = (trim(suExec("getnfsstatus")) === "true");
 	$afp_running = (trim(suExec("getafpstatus")) === "true");
 	if (!$tftp_running)
 	{
 		echo "<div class=\"alert alert-danger\">ERROR: TFTP is not running, restart NetBoot</div>";
 	}
-	/* if (!$nfs_running)
+	if (!$nfs_running)
 	{
 		echo "<div class=\"alert alert-danger\">ERROR: NFS is not running, restart NetBoot</div>";
-	} */
+	}
 	if (!$afp_running)
 	{
 		echo "<div class=\"alert alert-warning\">WARNING: AFP is not running, diskless will be unavailable</div>";
@@ -170,13 +164,6 @@ function enableButton(id, enable)
 	document.getElementById(id).disabled = !enable;
 }
 
-function validateName()
-{
-	var validName = /^[A-Za-z0-9._ +\-]{1,256}$/.test(document.getElementById("netbootName").value) || document.getElementById("netbootName").value == "";
-	showErr("netbootName", validName);
-	enableButton("changenetboot", validName);
-}
-
 function validateSubnet()
 {
 	var validSubnet = /^(([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\.){3}([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])$/.test(document.getElementById("subnet").value);
@@ -199,7 +186,7 @@ function validateSubnet()
 			<br>
 
 			<?php
-			if (getNetBootStatus())
+			if (!isset($_POST['disablenetboot']) && getNetBootStatus())
 			{
 				echo "<div class=\"alert alert-success alert-with-button\">
 						<span>Enabled</span>
@@ -223,53 +210,53 @@ function validateSubnet()
 
 			<br><br>
 
-			<div class="panel panel-default">
+			<div class="table-responsive panel panel-default">
 				<div class="panel-heading">
-					<strong>NetBoot Image and Name</strong>
+					<strong>NetBoot Images</strong>
 				</div>
-				<div class="panel-body">
-					<div class="input-group">
-						<div class="input-group-addon no-background">Image</div>
-						<span class="description">NetBoot image that computers boot to</span>
-						<select name="NetBootImage" id="NetBootImage" class="form-control input-sm" onChange="javascript:ajaxPost('ajax.php?service=NetBoot', 'NetBootImage='+this.value);">
-							<?php
-							$nbidircontents = scandir($netbootimgdir);
-							$curimg = $conf->getSetting("netbootimage");
+				<table class="table table-striped table-bordered table-condensed">
+					<thead>
+						<tr>
+							<th>Enable</th>
+							<th>Image</th>
+							<th>Name</th>
+							<th>Index</th>
+							<th>Type</th>
+						</tr>
+					</thead>
+					<tbody>
+						<?php
+						$nbidircontents = scandir($netbootimgdir);
+						$curimg = $conf->getSetting("netbootimage");
+						foreach($nbidircontents as $item)
+						{
 							$i = 0;
-							foreach($nbidircontents as $item)
+							if ($item != "." && $item != ".." && is_dir($netbootimgdir.$item) && file_exists($netbootimgdir.$item."/i386/booter"))
 							{
-								if ($item != "." && $item != ".." && is_dir($netbootimgdir.$item) && file_exists($netbootimgdir.$item."/i386/booter"))
-								{
-									?>
-									<option value="<?php echo $item?>" <?php echo ($curimg == $item ? "selected=\"selected\"" : "")?>><?php echo $item?></option>
-									<?php
-								}
-								$i++;
+								$Name = trim(suExec("getNBIproperty ".$item." Name"));
+								if ($Name == "") { $Name = str_replace(".nbi", "" , $item); }
+								$Index = trim(suExec("getNBIproperty ".$item." Index"));
+								if ($Index == "") { $Index = "526"; }
+								$Type = trim(suExec("getNBIproperty ".$item." Type"));
+								if ($Type == "") { $Type = "HTTP"; }
+								?>
+						<tr>
+							<td><input type="radio" name="Enabled[<?php echo $i; ?>]" id="Enabled[<?php echo $i; ?>]" value="<?php echo $item ?>" <?php if ($curimg == $item) { echo "checked=\"checked\""; $NetBootImage = $item; } ?> onChange="document.getElementById('NetBootImage').value = this.value; javascript:ajaxPost('ajax.php?service=NetBoot', 'NetBootImage='+this.value);"/></td>
+							<td><a href="managenbi.php?image=<?php echo $item?>"><?php echo $item?></a></td>
+							<td><?php echo $Name ?></td>
+							<td><?php echo $Index ?></td>
+							<td><?php echo $Type ?></td>
+						</tr>
+								<?php
 							}
-
-							if ($i == 0)
-							{
-								echo "<option value=\"\">---</option>\n";
-							}
-
-							?>
-						</select>
-					</div>
-
-					<br>
-
-					<div class="input-group">
-						<div class="input-group-addon no-background">Name</div>
-						<span class="description">(Optional) NetBoot name to appear on receiving boot devices. Defaults to the .nbi folder name</span>
-						<input type="text" name="netbootName" id="netbootName" class="form-control input-sm" value="<?php echo $conf->getSetting("netbootname")?>" onKeyUp="validateName();" onChange="validateName();" />
-					</div>
-				</div>
-
-				<div class="panel-footer">
-					<input type="submit" name="changenetboot" id="changenetboot" class="btn btn-primary btn-sm" value="Change" disabled="disabled" />
-				</div>
-
+							$i++;
+						}
+						?>
+					</tbody>
+				</table>
 			</div>
+
+			<input type="hidden" id="NetBootImage" name="NetBootImage" value="<?php echo $NetBootImage; ?>" />
 
 			<div class="panel panel-default">
 				<div class="panel-heading">
@@ -281,14 +268,14 @@ function validateSubnet()
 					<div class="input-group">
 						<div class="input-group-addon no-background">Subnet</div>
 						<span class="description">One of the subnets must include the IP address of the NetBoot server</span>
-						<input type="text" name="subnet" id="subnet" class="form-control input-sm" value="<?php if (!array_key_exists($currentSubnet." ".$currentNetmask, $conf->getSubnets())) { echo $currentSubnet; } ?>" onKeyUp="validateSubnet();" onChange="validateSubnet();" />
+						<input type="text" name="subnet" id="subnet" class="form-control input-sm" value="<?php if (!array_key_exists($currentSubnet." ".$currentNetmask, $conf->getSubnets())) { echo $currentSubnet; } ?>" onClick="validateSubnet();" onKeyUp="validateSubnet();" onChange="validateSubnet();" />
 					</div>
 
 					<br>
 
 					<div class="input-group">
 						<div class="input-group-addon no-background">Netmask</div>
-						<input class="form-control round-corners input-sm" type="text" name="netmask" id="netmask" value="<?php if (!array_key_exists($currentSubnet." ".$currentNetmask, $conf->getSubnets())) { echo $currentNetmask; } ?>" onKeyUp="validateSubnet();" onChange="validateSubnet();" />
+						<input class="form-control round-corners input-sm" type="text" name="netmask" id="netmask" value="<?php if (!array_key_exists($currentSubnet." ".$currentNetmask, $conf->getSubnets())) { echo $currentNetmask; } ?>" onClick="validateSubnet();" onKeyUp="validateSubnet();" onChange="validateSubnet();" />
 					</div>
 
 				</div>
@@ -312,7 +299,7 @@ function validateSubnet()
 						<tr class="<?php echo ($key % 2 == 0 ? "object0" : "object1")?>">
 							<td><?php echo $value['subnet']?></td>
 							<td><?php echo $value['netmask']?></td>
-							<td><a href="netBoot.php?service=NetBoot&deleteSubnet=<?php echo urlencode($value['subnet'])?>&deleteNetmask=<?php echo urlencode($value['netmask'])?>">Delete</a>
+							<td><a href="netBoot.php?service=NetBoot&deleteSubnet=<?php echo urlencode($value['subnet'])?>&deleteNetmask=<?php echo urlencode($value['netmask'])?>">Delete</a></td>
 						</tr>
 						<?php } ?>
 					</tbody>